Made more changes
