+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
 
         <activity
-            android:name=".MainActivity"
+            android:name=".OpenAnimation"
             android:exported="true">
 
             <intent-filter>
@@ -18,7 +18,7 @@
 
         </activity>
 
-        <activity android:name=".HomeActivity" android:exported="true" />
+        <activity android:name=".MainActivity" android:exported="true" />
 
 
     </application>
